+++ lizard_progress/views/views.py
@@ -564,8 +564,7 @@ class InlineMapViewNew(View):
         chreqs = Request.objects.filter(activity__in=activities).distinct()
         layers = {}
 
-        cursor = connection.cursor()
-        try:
+        with connection.cursor() as cursor:
             # Activities
             for a in activities:
                 q = """select json_build_object(
@@ -645,8 +644,7 @@ class InlineMapViewNew(View):
                 cursor.execute(q)
                 features = [json.loads(r[0]) for r in cursor.fetchall()]
                 layers['OoI'] = geojson.FeatureCollection(features)
-        finally:
-            cursor.close()
+            
 
         res = json.dumps(layers)
         return (res)