+++ CHANGES.rst
@@ -10,7 +10,7 @@ Changelog of lizard-progress
 4.2.4 (2018-07-30)
 ------------------
 
-- Hotfix connection.cursor for python 2.7 w/o DebugToolbar.
+- Nothing changed yet.
 
 
 4.2.3 (2018-07-30)